% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{aic.fisher_mle}
\alias{aic.fisher_mle}
\title{Compute AIC for fisher_mle}
\usage{
\method{aic}{fisher_mle}(x, ...)
}
\arguments{
\item{x}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
AIC value
}
\description{
Returns the Akaike Information Criterion:
AIC = -2 * logL + 2 * k
where k is the number of parameters.
}
