% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{algebraic.mle-reexports}
\alias{algebraic.mle-reexports}
\alias{se}
\alias{bias}
\alias{aic}
\alias{loglik_val}
\alias{score_val}
\alias{sampler}
\title{Generics re-exported from algebraic.mle}
\description{
These generics are defined in \pkg{algebraic.mle} and re-exported
here so that users of \pkg{likelihood.model} can access them
without explicitly loading \pkg{algebraic.mle}.
}
\details{
\describe{
\item{\code{\link[algebraic.mle]{se}}}{Standard errors of parameter estimates}
\item{\code{\link[algebraic.mle]{bias}}}{Bias of parameter estimates}
\item{\code{\link[algebraic.mle]{aic}}}{Akaike Information Criterion}
\item{\code{\link[algebraic.mle]{loglik_val}}}{Log-likelihood value at the MLE}
\item{\code{\link[algebraic.mle]{score_val}}}{Score vector at the MLE}
\item{\code{\link[algebraic.mle]{sampler}}}{Sampling distribution of the estimator}
}
}
