% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-exponential.R
\name{assumptions.exponential_lifetime}
\alias{assumptions.exponential_lifetime}
\title{Assumptions for exponential_lifetime}
\usage{
\method{assumptions}{exponential_lifetime}(model, ...)
}
\arguments{
\item{model}{An exponential_lifetime model}

\item{...}{Additional arguments (ignored)}
}
\value{
Character vector of assumptions
}
\description{
Assumptions for exponential_lifetime
}
