% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-contr.R
\name{assumptions.likelihood_contr_model}
\alias{assumptions.likelihood_contr_model}
\title{Retrieve the assumptions in the likelihood contributions model.}
\usage{
\method{assumptions}{likelihood_contr_model}(model, ...)
}
\arguments{
\item{model}{The likelihood contribution model}

\item{...}{Additional arguments}
}
\value{
A list of assumptions
}
\description{
Retrieve the assumptions in the likelihood contributions model.
}
