% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_exact_weibull.R
\name{assumptions.likelihood_exact_weibull}
\alias{assumptions.likelihood_exact_weibull}
\title{List the assumptions made by the model.}
\usage{
\method{assumptions}{likelihood_exact_weibull}(model, ...)
}
\arguments{
\item{model}{The weibull likelihood model}
}
\description{
Since this is a very simple model that assumes only one
observation type, the assumptions are simply "independent"
and "identically distributed".
}
