% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-generics.R
\name{assumptions}
\alias{assumptions}
\title{Retrieve the assumptions the likelihood model makes about the data.}
\usage{
assumptions(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
A list of assumptions
}
\description{
Retrieve the assumptions the likelihood model makes about the data.
}
