% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-weibull.R
\name{assumptions.weibull_uncensored}
\alias{assumptions.weibull_uncensored}
\title{List the assumptions made by the model.}
\usage{
\method{assumptions}{weibull_uncensored}(model, ...)
}
\arguments{
\item{model}{The weibull likelihood model}

\item{...}{Additional arguments (ignored)}
}
\value{
Character vector of assumptions
}
\description{
Since this is a very simple model that assumes only one
observation type, the assumptions are simply "independent"
and "identically distributed".
}
