% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{bic}
\alias{bic}
\alias{bic.fisher_mle}
\title{Compute BIC}
\usage{
bic(x, ...)

\method{bic}{fisher_mle}(x, ...)
}
\arguments{
\item{x}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
BIC value
}
\description{
Returns the Bayesian Information Criterion:
BIC = -2 * logL + k * log(n)
where k is the number of parameters and n is the sample size.
}
