% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model_boot.R
\name{bootstrap.likelihood_model}
\alias{bootstrap.likelihood_model}
\title{Bootstrap a likelihood model}
\usage{
\method{bootstrap}{likelihood_model}(
  model,
  par = NULL,
  R = 999,
  algo = stats::optim,
  method_seq = c("SANN", "Nelder-Mead"),
  nthreads = NULL,
  ...
)

\method{bootstrap}{likelihood_model}(
  model,
  par = NULL,
  R = 999,
  algo = stats::optim,
  method_seq = c("SANN", "Nelder-Mead"),
  nthreads = NULL,
  ...
)
}
\arguments{
\item{model}{The likelihood model}

\item{par}{The starting parameter values}

\item{R}{The number of bootstrap samples to take}

\item{algo}{The algorithm to use for the MLE solver. It should
look like the \code{stats::optim} function, i.e., take similar arguments
(if you don't do anything with them in your algorithm, just accept them
as \code{...} and don't do anything with them) and, especially, return a list with
the same named values (it may also return more). Defaults to \code{stats::optim}.}

\item{method_seq}{The sequence of optimization methods to use. By default,
we use SANN and then Nelder-Mead.}

\item{nthreads}{The number of threads to use for parallelization}

\item{...}{Additional arguments to pass into the likelihood model's}
}
\value{
A function that returns an empirical sampling distribution of an
MLE.

A function that returns an bootstrapped sampling distribution of an
MLE.
}
\description{
In other words, we treat the data as an empirical distribution and
sample from it to get a new dataset, then we fit the model to that dataset
and return the MLE. We do this R times and return the R MLEs.
}
