% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model_boot.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap a model. The statistics that are bootstrapped are the
parameters of the model, or other salient characteristics that
are important to the model. In a likelihood model, for instance,
the parameters are bootstrapped.}
\usage{
bootstrap(model, ...)
}
\arguments{
\item{model}{A likelihood model}

\item{...}{Additional arguments to \code{statistic}}
}
\value{
A function that returns an empirical sampling distribution of an
MLE.
}
\description{
Bootstrap a model. The statistics that are bootstrapped are the
parameters of the model, or other salient characteristics that
are important to the model. In a likelihood model, for instance,
the parameters are bootstrapped.
}
