% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{coef.fisher_mle}
\alias{coef.fisher_mle}
\title{Extract coefficients from fisher_mle object}
\usage{
\method{coef}{fisher_mle}(object, ...)
}
\arguments{
\item{object}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
Named numeric vector of parameter estimates
}
\description{
Extract coefficients from fisher_mle object
}
