% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-exponential.R
\name{.exponential_sufficient_stats}
\alias{.exponential_sufficient_stats}
\title{Compute sufficient statistics for the exponential model}
\usage{
.exponential_sufficient_stats(df, model)
}
\arguments{
\item{df}{Data frame}

\item{model}{An exponential_lifetime model}
}
\value{
List with d (number of exact obs), total_time (sum of all times),
n (total observations)
}
\description{
Compute sufficient statistics for the exponential model
}
\keyword{internal}
