% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-contr.R
\name{.new_contr_cache}
\alias{.new_contr_cache}
\title{Create a new cache environment for likelihood_contr_model wrappers.}
\usage{
.new_contr_cache()
}
\value{
An environment with df_ref, dfs, and fns slots
}
\description{
Create a new cache environment for likelihood_contr_model wrappers.
}
\keyword{internal}
