% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_likelihood.R
\name{fim.likelihood_exact_normal_model}
\alias{fim.likelihood_exact_normal_model}
\title{FIM function generator for the normal distribution with
exact observations.}
\usage{
\method{fim}{likelihood_exact_normal_model}(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments (not used)}
}
\value{
The FIM function
}
\description{
FIM function generator for the normal distribution with
exact observations.
}
