% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{fim}
\alias{fim}
\title{Fisher information matrix method}
\usage{
fim(model, par, ...)
}
\arguments{
\item{model}{likelihood model}

\item{par}{true parameters}

\item{...}{additional arguments}
}
\value{
FIM
}
\description{
Compute the Fisher information matrix (FIM), which is
an expectation over the data-generating process (DGP):
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{FIM(theta|x) = E[-loglik_hessian(x)(data,theta)] / n
FIM(theta|x) = E[score(x)(data,theta) \%*\% t(score(x)(data,theta))] / n
}\if{html}{\out{</div>}}

where \code{x} is the likelihood model (say, exact failure times for
exponentially distributed lifetimes), \code{theta} is the parameters
of the likelihood model, \code{loglik_hessian} computes the hessian
of the log-likelihood function on the \code{data} (which is a random
quantity that we take the expectation over), and \code{score} is
the score function, the gradient of the log-likelhood function, and
\code{n} is the number of observations in each \code{data}. The FIM is
the negative of the expected hessian of the log-likelihood function
(or the expected outer product of the score function).

The FIM is a function of the parameters, and is used to compute
the standard errors of the parameters. The FIM is also used to
compute the covariance matrix of the parameters, which is used
to compute the standard errors of the parameters.

The FIM is also used to compute the Cramer-Rao lower bound (CRLB),
which is the inverse of the FIM. The CRLB is the minimum variance
of an unbiased estimator of the parameters. The CRLB is used to
compute the asymptotic relative efficiency (ARE) of an estimator
of the parameters, which is the ratio of the variance of the
estimator to the CRLB.
}
