% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_likelihood.R
\name{fit.likelihood_exact_normal}
\alias{fit.likelihood_exact_normal}
\title{Customized MLE fit for the exact normal likelihood model.}
\usage{
\method{fit}{likelihood_exact_normal}(model, keep_obs = TRUE, ...)
}
\arguments{
\item{model}{normal likelihood model}

\item{keep_obs}{whether to store observations in \code{mle_exact_normal} object}
}
\value{
an \code{mle} object of class \code{mle_exact_normal}
}
\description{
Customized MLE fit for the exact normal likelihood model.
}
