% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{fit}
\alias{fit}
\title{Generic method for maximum likelihood estimation (MLE)
according to a likelihood model.}
\usage{
fit(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\description{
Generic method for maximum likelihood estimation (MLE)
according to a likelihood model.
}
