% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-exponential.R
\name{hess_loglik.exponential_lifetime}
\alias{hess_loglik.exponential_lifetime}
\title{Hessian of the log-likelihood for exponential_lifetime}
\usage{
\method{hess_loglik}{exponential_lifetime}(model, ...)
}
\arguments{
\item{model}{An exponential_lifetime model}

\item{...}{Additional arguments (ignored)}
}
\value{
A function(df, par, ...) computing the Hessian matrix
}
\description{
Returns a function computing the 1x1 Hessian: -d/lambda^2.
}
