% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_contr_model.R
\name{hess_loglik.likelihood_contr_model}
\alias{hess_loglik.likelihood_contr_model}
\title{Hessian of log-likelihood method for likelihood_contr_model}
\usage{
\method{hess_loglik}{likelihood_contr_model}(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
A function to compute the hessian of the log-likelihood given a
data frame and parameters
}
\description{
This method returns the hessian of the log-likelihood function for a
likelihood_contr_model
}
