% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_likelihood.R
\name{hess_loglik.likelihood_exact_normal}
\alias{hess_loglik.likelihood_exact_normal}
\title{Fisher information matrix generator given data \code{x} for the normal
distribution}
\usage{
\method{hess_loglik}{likelihood_exact_normal}(model, ...)
}
\arguments{
\item{...}{additional arguments (not used)}

\item{x}{data (simple random sample)}

\item{observed}{whether to return the observed fisher information, default is
\code{TRUE}}
}
\description{
Fisher information matrix generator given data \code{x} for the normal
distribution
}
