% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_exact_weibull.R
\name{hess_loglik.likelihood_exact_weibull}
\alias{hess_loglik.likelihood_exact_weibull}
\title{Hessian of the log-likelihood function generator for the weibull
likelihood model.}
\usage{
\method{hess_loglik}{likelihood_exact_weibull}(model, ...)
}
\arguments{
\item{model}{The weibull likelihood model}
}
\description{
Hessian of the log-likelihood function generator for the weibull
likelihood model.
}
