% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{hess_loglik}
\alias{hess_loglik}
\title{Hessian of log-likelihood method}
\usage{
hess_loglik(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
A function to compute the hessian of the log-likelihood given a
data frame and parameters
}
\description{
This function returns the hessian of the log-likelihood function of a model
}
