% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-weibull.R
\name{hess_loglik.weibull_uncensored}
\alias{hess_loglik.weibull_uncensored}
\title{Hessian of the log-likelihood function generator for the Weibull
uncensored model.}
\usage{
\method{hess_loglik}{weibull_uncensored}(model, ...)
}
\arguments{
\item{model}{The weibull likelihood model}

\item{...}{Additional arguments (ignored)}
}
\value{
A function to compute the Hessian of the log-likelihood given
a data frame and parameters
}
\description{
Hessian of the log-likelihood function generator for the Weibull
uncensored model.
}
