% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-generics.R
\name{is_likelihood_model}
\alias{is_likelihood_model}
\title{Check if an object is a likelihood model}
\usage{
is_likelihood_model(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical indicating whether x is a likelihood model
}
\description{
Tests whether an object satisfies the \code{likelihood_model} concept by
checking if \code{"likelihood_model"} is in its class hierarchy. To be a
likelihood model, at a minimum the object must implement \code{\link[=loglik]{loglik()}}.
For optimal results, it may also provide \code{\link[=score]{score()}} and \code{\link[=hess_loglik]{hess_loglik()}}.
}
