% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{lik.likelihood_model}
\alias{lik.likelihood_model}
\title{Likelihood method}
\usage{
\method{lik}{likelihood_model}(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments to pass into the \code{loglik} function}
}
\value{
A likelihood function to compute the likelihood given a data frame
}
\description{
This function returns the likelihood function of a model.
}
