% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{lik}
\alias{lik}
\title{Likelihood method}
\usage{
lik(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
A likelihood function to compute the likelihood given a data frame
and parameters.
}
\description{
This function returns the likelihood function of a model.
}
