% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_likelihood.R
\name{likelihood_exact_normal}
\alias{likelihood_exact_normal}
\title{Exact Weibull likelihood model}
\usage{
likelihood_exact_normal(ob_col)
}
\arguments{
\item{ob_col}{The name of the column in a data frame that
contains the observations.}
}
\description{
A likelihood model for observations that are exact
observations of a random variable that follows a Normal
distribution. The model assumes that the observations are
independent and identically distributed (i.i.d.).

This model defines all the relevant functions to satisfy
being a likelihood model. It explicitly provides the following
functions:
\itemize{
\item \code{loglik.likelihood_exact_normal}: generates the log-likelihood
function for the model.
\item \code{score.likelihood_exact_normal}: generates the score function
for the model.
\item \code{hess_loglik.likelihood_exact_normal}: generates the Hessian
of the log-likelihood for the model.
}

Note: This model may be used as a contribution in the
\code{likelihood_contr_model} if a more complicated likelihood model
is needed. We define this simple model mostly for illustrative
purposes.
}
