% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{likelihood_model}
\alias{likelihood_model}
\title{likelihood_model class}
\description{
likelihood_model class

likelihood_model class
}
\details{
This class encapsulates all necessary parts of a likelihood model.
It provides methods for computing the log-likelihood, score, and Hessian
of the model. It also allows for different functions depending on the
observation type.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{logliks}}{list of functions for computing log-likelihoods}

\item{\code{scores}}{list of functions for computing scores}

\item{\code{hess_logliks}}{list of functions for computing Hessians}

\item{\code{obs_type}}{function that determines observation type}

\item{\code{assumptions}}{list of assumptions made by the model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-likelihood_model-new}{\code{likelihood_model$new()}}
\item \href{#method-likelihood_model-dispatch}{\code{likelihood_model$dispatch()}}
\item \href{#method-likelihood_model-loglik}{\code{likelihood_model$loglik()}}
\item \href{#method-likelihood_model-score}{\code{likelihood_model$score()}}
\item \href{#method-likelihood_model-hess_loglik}{\code{likelihood_model$hess_loglik()}}
\item \href{#method-likelihood_model-clone}{\code{likelihood_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-likelihood_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-likelihood_model-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{likelihood_model$new(
  logliks,
  obs_type,
  scores = NULL,
  hess_logliks = NULL,
  assumptions = c("iid")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{logliks}}{list of functions for computing log-likelihoods}

\item{\code{obs_type}}{function that determines observation type}

\item{\code{scores}}{list of functions for computing scores, default is
NULL (finite difference method is used)}

\item{\code{hess_logliks}}{list of functions for computing Hessians,
default is NULL (finite difference method is used)}

\item{\code{assumptions}}{list of assumptions made by the model, default is
c("iid""), which assumes iid observations (this assumption is always
made for this class, which is why we can sum the log-likelihood
contributions)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{likelihood_model} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-likelihood_model-dispatch"></a>}}
\if{latex}{\out{\hypertarget{method-likelihood_model-dispatch}{}}}
\subsection{Method \code{dispatch()}}{
Generic dispatcher for calculating total log-likelihood, score,
and Hessian.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{likelihood_model$dispatch(df, par, methods, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe for computation}

\item{\code{par}}{parameters for computation}

\item{\code{methods}}{methods to be dispatched}

\item{\code{...}}{additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The sum of the dispatched methods
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-likelihood_model-loglik"></a>}}
\if{latex}{\out{\hypertarget{method-likelihood_model-loglik}{}}}
\subsection{Method \code{loglik()}}{
Computes the log-likelihood of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{likelihood_model$loglik(df, par, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe for computation}

\item{\code{par}}{parameters for computation}

\item{\code{...}}{additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The total log-likelihood
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-likelihood_model-score"></a>}}
\if{latex}{\out{\hypertarget{method-likelihood_model-score}{}}}
\subsection{Method \code{score()}}{
Computes the score of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{likelihood_model$score(df, par, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe for computation}

\item{\code{par}}{parameters for computation}

\item{\code{...}}{additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The total score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-likelihood_model-hess_loglik"></a>}}
\if{latex}{\out{\hypertarget{method-likelihood_model-hess_loglik}{}}}
\subsection{Method \code{hess_loglik()}}{
Computes the Hessian of the log-likelihood.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{likelihood_model$hess_loglik(df, par, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe for computation}

\item{\code{par}}{parameters for computation}

\item{\code{...}}{additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The Hessian of the log-likelihood
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-likelihood_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-likelihood_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{likelihood_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
