% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model_boot.R
\name{likelihood_model_boot}
\alias{likelihood_model_boot}
\title{Likelihood model bootstrapping}
\description{
Bootstrap a likelihood model to get standard errors and confidence intervals
for the parameters when the asymptotic distribution may not kick in.

When you do the bootstrap on the model, it returns a boot object
(see \code{boot}), but specialized for maximum likelihood estimation, and so
we implement the \code{mle} API for it and assign it the class \code{mle_boot}.

For many methods, it just calls the boot method, e.g., CIs can be computed
with \code{boot.ci}, but we also provide methods for implementnig other kinds
of methods appropriate to \code{mle} objects, like \code{vcov} (variance-covariance
matrix) and such.
}
