% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-exponential.R
\name{loglik.exponential_lifetime}
\alias{loglik.exponential_lifetime}
\title{Log-likelihood for exponential_lifetime}
\usage{
\method{loglik}{exponential_lifetime}(model, ...)
}
\arguments{
\item{model}{An exponential_lifetime model}

\item{...}{Additional arguments (ignored)}
}
\value{
A function(df, par, ...) computing the log-likelihood
}
\description{
Returns a function computing ell(lambda) = d * log(lambda) - lambda * T.
}
