% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{logLik.fisher_mle}
\alias{logLik.fisher_mle}
\title{Extract log-likelihood from fisher_mle object}
\usage{
\method{logLik}{fisher_mle}(object, ...)
}
\arguments{
\item{object}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
A logLik object
}
\description{
Extract log-likelihood from fisher_mle object
}
