% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_likelihood.R
\name{loglik.likelihood_exact_normal}
\alias{loglik.likelihood_exact_normal}
\title{Log-likelihood function generator for the exact Weibull likelihood model.}
\usage{
\method{loglik}{likelihood_exact_normal}(model, ...)
}
\arguments{
\item{model}{The weibull likelihood model}

\item{...}{Additional arguments (not used)}
}
\value{
A function to compute the log-likelihood given a data frame and
parameters.
}
\description{
Log-likelihood function generator for the exact Weibull likelihood model.
}
