% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_name.R
\name{loglik.likelihood_name_model}
\alias{loglik.likelihood_name_model}
\title{Log-likelihood function for named distribution models}
\usage{
\method{loglik}{likelihood_name_model}(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments (not used)}
}
\value{
A function(df, par, ...) that computes the log-likelihood
}
\description{
Log-likelihood function for named distribution models
}
