% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-weibull.R
\name{loglik.weibull_uncensored}
\alias{loglik.weibull_uncensored}
\title{Log-likelihood function generator for the Weibull uncensored model.}
\usage{
\method{loglik}{weibull_uncensored}(model, ...)
}
\arguments{
\item{model}{The weibull likelihood model}

\item{...}{Additional arguments (not used)}
}
\value{
A function to compute the log-likelihood given a data frame and
parameters.
}
\description{
Log-likelihood function generator for the Weibull uncensored model.
}
