% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{loglik_val.fisher_mle}
\alias{loglik_val.fisher_mle}
\title{Extract log-likelihood value from fisher_mle}
\usage{
\method{loglik_val}{fisher_mle}(x, ...)
}
\arguments{
\item{x}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
The log-likelihood value at the MLE
}
\description{
Extract log-likelihood value from fisher_mle
}
