% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-lrt.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test}
\usage{
lrt(null, alt, data, null_par, alt_par, dof = NULL, ...)
}
\arguments{
\item{null}{the likelihood model for the simpler (null) hypothesis nested
within the alternative model}

\item{alt}{the likelihood model for the more complicated (alternative)
hypothesis}

\item{data}{a data frame}

\item{null_par}{parameter values under the null model}

\item{alt_par}{parameter values under the alternative model}

\item{dof}{degrees of freedom (computed automatically if NULL)}

\item{...}{additional arguments passed to loglik}
}
\value{
likelihood ratio test
}
\description{
Likelihood ratio test
}
