% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{mle.likelihood_model}
\alias{mle.likelihood_model}
\title{Default MLE solver for subclasses of \code{likelihood_model}.}
\usage{
\method{mle}{likelihood_model}(model, data, par0, control = list(), ...)
}
\arguments{
\item{model}{The likelihood model}

\item{data}{The data}

\item{par0}{The initial guess for the parameters}

\item{control}{Control parameters}

\item{...}{Additional arguments to pass into the likelihood model's
\code{loglik} and \code{score} methods}
}
\value{
The MLE of the parameters
}
\description{
Note that \code{likelihood_model} is not a class, but a concept,
that other likelihood models implement. They should add
\code{likelihood_model} to their class definition, and then
they can use this function to compute the MLE.
}
\details{
This function uses the \code{optim} function to find the MLE of the
parameters of a likelihood model. It uses the \code{loglik} and \code{score}
methods to compute the log-likelihood and score function, respectively.
}
