% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{mle}
\alias{mle}
\title{Generic method for maximum likelihood estimation (MLE)
according to a likelihood model.}
\usage{
mle(model, data, par0, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{data}{The data}

\item{par0}{The initial guess for the parameters}

\item{...}{Additional arguments}
}
\description{
Generic method for maximum likelihood estimation (MLE)
according to a likelihood model.
}
