% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{nobs.fisher_mle}
\alias{nobs.fisher_mle}
\title{Extract number of observations from fisher_mle object}
\usage{
\method{nobs}{fisher_mle}(object, ...)
}
\arguments{
\item{object}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
Number of observations
}
\description{
Extract number of observations from fisher_mle object
}
