% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{nparams.fisher_mle}
\alias{nparams.fisher_mle}
\title{Number of parameters in fisher_mle}
\usage{
\method{nparams}{fisher_mle}(x)
}
\arguments{
\item{x}{A fisher_mle object}
}
\value{
Integer count of parameters
}
\description{
Number of parameters in fisher_mle
}
