% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{observed_fim.fisher_mle}
\alias{observed_fim.fisher_mle}
\title{Observed Fisher information matrix from fisher_mle}
\usage{
\method{observed_fim}{fisher_mle}(x, ...)
}
\arguments{
\item{x}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
A matrix, or NULL if the Hessian was not computed
}
\description{
Returns the negative Hessian of the log-likelihood evaluated at the
MLE, which estimates the Fisher information matrix.
}
