% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{observed_info.likelihood_model}
\alias{observed_info.likelihood_model}
\title{Default observed information method}
\usage{
\method{observed_info}{likelihood_model}(model, ...)
}
\arguments{
\item{model}{A likelihood model}

\item{...}{Additional arguments passed to hess_loglik}
}
\value{
Function that takes (df, par, ...) and returns observed information matrix
}
\description{
Returns a function that computes -hess_loglik(df, par).
}
