% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{params.fisher_mle}
\alias{params.fisher_mle}
\title{Extract parameter estimates from fisher_mle}
\usage{
\method{params}{fisher_mle}(x)
}
\arguments{
\item{x}{A fisher_mle object}
}
\value{
Named numeric vector of parameter estimates
}
\description{
Extract parameter estimates from fisher_mle
}
