% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{print.fisher_boot}
\alias{print.fisher_boot}
\title{Print fisher_boot object}
\usage{
\method{print}{fisher_boot}(x, ...)
}
\arguments{
\item{x}{A fisher_boot object}

\item{...}{Additional arguments (ignored)}
}
\value{
The fisher_boot object, invisibly
}
\description{
Print fisher_boot object
}
