% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{print.fisher_mle}
\alias{print.fisher_mle}
\title{Print fisher_mle object}
\usage{
\method{print}{fisher_mle}(x, ...)
}
\arguments{
\item{x}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
The fisher_mle object, invisibly
}
\description{
Print fisher_mle object
}
