% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisherian.R
\name{print.likelihood_interval}
\alias{print.likelihood_interval}
\title{Print likelihood interval}
\usage{
\method{print}{likelihood_interval}(x, ...)
}
\arguments{
\item{x}{A likelihood_interval object}

\item{...}{Additional arguments (ignored)}
}
\value{
The likelihood_interval object, invisibly
}
\description{
Print likelihood interval
}
