% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{print.likelihood_model}
\alias{print.likelihood_model}
\title{Print method for likelihood models}
\usage{
\method{print}{likelihood_model}(x, ...)
}
\arguments{
\item{x}{A likelihood model}

\item{...}{Additional arguments}
}
\description{
Print method for likelihood models
}
