% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{print.summary_fisher_mle}
\alias{print.summary_fisher_mle}
\title{Print summary of fisher_mle}
\usage{
\method{print}{summary_fisher_mle}(x, ...)
}
\arguments{
\item{x}{A summary_fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
The summary_fisher_mle object, invisibly
}
\description{
Print summary of fisher_mle
}
