% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{rdata}
\alias{rdata}
\title{Random data generation method}
\usage{
rdata(model, ...)
}
\arguments{
\item{model}{A likelihood model}

\item{...}{Additional arguments}
}
\value{
Function that takes (theta, n, ...) and returns a data frame
}
\description{
Returns a function that generates random data from the model's
data-generating process (DGP) at a given parameter value.
}
\details{
This is used by the default \code{fim} method for Monte Carlo estimation
of the Fisher information matrix.
}
