% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-exponential.R
\name{score.exponential_lifetime}
\alias{score.exponential_lifetime}
\title{Score for exponential_lifetime}
\usage{
\method{score}{exponential_lifetime}(model, ...)
}
\arguments{
\item{model}{An exponential_lifetime model}

\item{...}{Additional arguments (ignored)}
}
\value{
A function(df, par, ...) computing the score vector
}
\description{
Returns a function computing d/lambda - T.
}
