% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{score}
\alias{score}
\title{Score method}
\usage{
score(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
A function to compute the score given a data frame and parameters
}
\description{
This function returns the score function of a model
}
