% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-weibull.R
\name{score.weibull_uncensored}
\alias{score.weibull_uncensored}
\title{Score function generator for the Weibull uncensored model.}
\usage{
\method{score}{weibull_uncensored}(model, ...)
}
\arguments{
\item{model}{The weibull likelihood model}

\item{...}{Additional arguments (not used)}
}
\value{
A function to compute the score given a data frame and parameters
}
\description{
Score function generator for the Weibull uncensored model.
}
