% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{summary.fisher_mle}
\alias{summary.fisher_mle}
\title{Summarize fisher_mle object}
\usage{
\method{summary}{fisher_mle}(object, ...)
}
\arguments{
\item{object}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
A summary_fisher_mle object
}
\description{
Summarize fisher_mle object
}
