% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-fisher_mle.R
\name{vcov.fisher_mle}
\alias{vcov.fisher_mle}
\title{Extract variance-covariance matrix from fisher_mle object}
\usage{
\method{vcov}{fisher_mle}(object, ...)
}
\arguments{
\item{object}{A fisher_mle object}

\item{...}{Additional arguments (ignored)}
}
\value{
Variance-covariance matrix
}
\description{
Extract variance-covariance matrix from fisher_mle object
}
