% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-name.R
\name{assumptions.likelihood_name_model}
\alias{assumptions.likelihood_name_model}
\title{List the assumptions made by the model}
\usage{
\method{assumptions}{likelihood_name_model}(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
Character vector of assumptions
}
\description{
List the assumptions made by the model
}
