% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{fim.likelihood_model}
\alias{fim.likelihood_model}
\title{Default implementation of the Fisher information matrix (FIM)
for a likelihood model}
\usage{
\method{fim}{likelihood_model}(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments to pass into the score and hess_loglik function
generators.}
}
\value{
The empirical FIM function, which takes as arguments:
\itemize{
\item \code{par}: The parameter vector
\item \code{df}: a data frame of observations in the sample used to estimate the FIM. If
you can do a Monte-carlo simulation, a very large sample provides
a estimate of the FIM (MC integration of the expectation of the
outer product of the score function).
}
}
\description{
This is an empirical FIM, which is the average of the outer product
of the score function over a empirical sample (\code{data}) or a large simulated
sample. The empirical FIM is an estimate of the true FIM.
}
\details{
Another good estimate of the FIM is just the hessian of the log-likelihood
evaluated at the MLE.
}
