% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{fim.likelihood_model}
\alias{fim.likelihood_model}
\title{Default implementation of the Fisher information matrix (FIM)
for a likelihood model}
\usage{
\method{fim}{likelihood_model}(model, par, data, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{par}{The true parameters}

\item{data}{The data}

\item{...}{Additional arguments}
}
\value{
The empirical FIM
}
\description{
Since we do not assume a particular DGP, and often times it is
difficult to compute the FIM analytically, we use the empirical
FIM, which is the average of the outer product of the score
function over a empirical sample (\code{data}). The empirical FIM
is an estimate of the true FIM.
}
\details{
It has a number of uses, e.g., its inverse is the Cramer-Rao
lower bound (CRLB), which is the minimum variance of an unbiased
estimator of the parameters. The CRLB has a number of uses,
such as computing the asymptotic sampling distribution of the
maximum likelihood estimator (MLE) of the parameters for a specified
likelihood model, which is asymptotically normal with mean
equal to the true parameters and variance equal to the inverse
of the FIM. This may be used to compute standard errors of the
parameters, confidence intervals, and other quantities.

This is per-observation. If you want the total FIM, multiply by the
number of observations for a given data set.

If the true parameters are not known, then you can use an MLE
or some other value (say in a hypothesis test) for the parameters.
}
\examples{
# generate data
df <- data.frame(x = rnorm(10000, 0, 1))

# define likelihood model
model <- likelihood_contr_model$new(
  logliks = list(
    normal = function(row, par) dnorm(row$x, par[1], par[2], log = TRUE)
  ),
  obs_type = function(row) "normal"
)

# compute FIM
fim(model, c(0, 1), df)
}
