% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{fim.likelihood_model}
\alias{fim.likelihood_model}
\title{Compute an estimate of the Fisher information matrix (FIM) using MC
simulation.}
\usage{
\method{fim}{likelihood_model}(model, par, data, ...)
}
\arguments{
\item{model}{likelihood model}

\item{par}{true parameters}

\item{data}{a sample from the assumed DGP (larger sample size is better)}

\item{...}{additional arguments}
}
\value{
MC estimate of the FIM
}
\description{
This is per-observation. If you want the total FIM, multiply by the
number of observations in a sample from the DGP.
}
\examples{
# generate data
df <- data.frame(x = rnorm(10000, 0, 1))

# define likelihood model
model <- likelihood_model$new(
  logliks = list(
    normal = function(row, par) dnorm(row$x, par[1], par[2], log = TRUE)
  ),
  obs_type = function(row) "normal"
)

# compute FIM
fim(model, c(0, 1), df)
}
