% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fim}
\alias{fim}
\title{Compute the Fisher information matrix (FIM), which is
an expectation over the data-generating process (DGP) rather than a
particular observed sample.}
\usage{
fim(model, par, ...)
}
\arguments{
\item{model}{likelihood model}

\item{par}{true parameters}

\item{...}{additional arguments}
}
\value{
FIM
}
\description{
Compute the Fisher information matrix (FIM), which is
an expectation over the data-generating process (DGP) rather than a
particular observed sample.
}
