% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{hess_loglik.likelihood_model}
\alias{hess_loglik.likelihood_model}
\title{Default method to compute the hessian of the log-likelihood.}
\usage{
\method{hess_loglik}{likelihood_model}(model, control = list(), ...)
}
\arguments{
\item{model}{The likelihood model}

\item{control}{Custom arguments to pass to \code{numDeriv::hessian}.}

\item{...}{Additional arguments (to pass into \code{loglik})}
}
\description{
In case a \code{hess_loglik} method is not provided, this function will be
used.
It computes the hessian of the log-likelihood function using numerical
differentiation.
}
