% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{is_likelihood_model}
\alias{is_likelihood_model}
\title{Likelihood model}
\usage{
is_likelihood_model(x)
}
\description{
A general likelihood model framework. When an object \code{x} implements
the concept of \code{likelihood_model}, it may be used to compute the
standard errors of the parameters of the model, maximum likelihood
estimates, the asymptotic sampling distribution of the MLE, and
other quantities.
}
\details{
To satisfy the concept of \code{likelihood_model}, at a minimum the
object must implement ``loglik\verb{. For optimal results, it may also provide implementations for }score`, and `hess_loglik` methods.
}
