% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{likelihood_model}
\alias{likelihood_model}
\title{Likelihood model}
\description{
A general likelihood model framework. When a likelihood model implements
this concept, it may be used to compute the log-likelihood, score, hessian
of the log-likelihood, Fisher information matrix, and other quantities for
the model. This framework is used to compute the standard errors of the
parameters of the model, maximum likelihood estimates, the asymptotic
sampling distribution of the MLE, and other quantities.
}
