% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_name.R
\name{likelihood_name}
\alias{likelihood_name}
\title{This is a likelihood model generator based on the convention
used in R, where we have a name for a distribution,
like \code{normal}, \code{weibull}, \code{exp}, etc. and then
we have a set of functions that are named
\verb{p<name>}, \verb{d<name>}, \verb{q<name>}, \verb{r<name>}, etc.,
where the first letter is the first letter of the
distribution name. For instance, for the normal
distribution, we have \code{pnorm}, \code{dnorm}, \code{qnorm},
\code{rnorm}, respectively for the cdf, pdf, quantile
function, and random variate generator.}
\usage{
likelihood_name(dist_name, ob_col, censor_col)
}
\arguments{
\item{dist_name}{The name of the distribution.}

\item{ob_col}{The name of the column in a data frame that
contains the observations.}

\item{type_col}{The name of the column in a data frame that
contains the censoring type. The values in this column determine the type of
censoring.
\itemize{
\item \code{exact} or NA: exact observation
\item \code{left}: left-censored observation
\item \code{right}: right-censored observation
\item \code{interval}: interval-censored observation
}}
}
\description{
This function generates a likelihood model based on
a distribution name. The distribution name is
used to generate the log-likelihood function,
the score function, and the Hessian of the
log-likelihood function for either exact (pdf),
left-censored (cdf), right-censored (1 - cdf),
or interval-censored (cdf(b) - cdf(a)) observations.
}
