% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{loglik.likelihood_model}
\alias{loglik.likelihood_model}
\title{Log-likelihood method for likelihood_model}
\usage{
\method{loglik}{likelihood_model}(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
A function to compute the log-likelihood given a data frame and
parameters
}
\description{
This method returns the log-likelihood function for a likelihood_model
}
