% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood method}
\usage{
loglik(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
A function to compute the log-likelihood given a data frame and
parameters
}
\description{
This function returns the log-likelihood function of a model
}
