% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test}
\usage{
lrt(model1, model2, data, par1, par2, dof = NULL, ...)
}
\arguments{
\item{data}{the data which the likelihood models are fitted to}

\item{dof}{degrees of freedom}

\item{...}{additional arguments}

\item{null}{the null (simpler) likelihood model}

\item{alt}{the more complicated model}

\item{null_par}{parameters for null model}

\item{alt_par}{parameters for alternative model}
}
\value{
likelihood ratio test
}
\description{
Likelihood ratio test
}
