% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test}
\usage{
lrt(null, alt, data, null_par, alt_par, dof = NULL, ...)
}
\arguments{
\item{null}{the likelihood value from a simpler likelihood model nested
within the alternative model}

\item{alt}{the likelihood value from the more complicated model}

\item{data}{a data frame}

\item{dof}{degrees of freedom}

\item{...}{additional arguments}
}
\value{
likelihood ratio test
}
\description{
Likelihood ratio test
}
