% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test}
\usage{
lrt(ll1, ll2, data, par1, par2, df = NULL, ...)
}
\arguments{
\item{data}{data}

\item{par1}{parameters for first model}

\item{par2}{parameters for second model}

\item{...}{additional arguments}

\item{model1}{first likelihood model}

\item{model2}{second likelihood model}
}
\value{
likelihood ratio test
}
\description{
Likelihood ratio test
}
