% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-generics.R
\name{print.likelihood_model}
\alias{print.likelihood_model}
\title{Print method for likelihood models}
\usage{
\method{print}{likelihood_model}(x, show.loglik = FALSE, ...)
}
\arguments{
\item{x}{A likelihood model}

\item{show.loglik}{Logical; if TRUE, print the log-likelihood function}

\item{...}{Additional arguments (ignored)}
}
\value{
The likelihood model object, invisibly
}
\description{
Print method for likelihood models
}
