% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-contr.R
\name{score.likelihood_contr_model}
\alias{score.likelihood_contr_model}
\title{Score method for likelihood_contr_model}
\usage{
\method{score}{likelihood_contr_model}(model, ...)
}
\arguments{
\item{model}{The likelihood model}

\item{...}{Additional arguments}
}
\value{
A function to compute the score given a data frame and parameters
}
\description{
Returns a score function with the same caching strategy as
\code{\link[=loglik.likelihood_contr_model]{loglik.likelihood_contr_model()}}.
}
