% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_exact_weibull.R
\name{score.likelihood_exact_weibull}
\alias{score.likelihood_exact_weibull}
\title{Score function generator for the exact Weibull likelihood model.}
\usage{
\method{score}{likelihood_exact_weibull}(model, ...)
}
\arguments{
\item{model}{The weibull likelihood model}

\item{...}{Additional arguments (not used)}
}
\description{
Score function generator for the exact Weibull likelihood model.
}
