% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_model.R
\name{score.likelihood_model}
\alias{score.likelihood_model}
\title{Default score method}
\usage{
\method{score}{likelihood_model}(model, control = list(), ...)
}
\arguments{
\item{model}{The likelihood model}

\item{control}{Custom arguments to pass to \code{numDeriv::grad}.}

\item{...}{Additional arguments (to pass into \code{loglik})}
}
\description{
In case a \code{score} method is not provided, this function will be used.
It computes the score by numerical differentiation of the log-likelihood.
}
